LED testing added to testing code + light reformatting
You know, it becomes really difficult to test something when pins on a board are reversed. Notes added to steering wheel board and code reflect changes that need to implemented for the next iteration of the board. Also, slightly reformatted some of the comment block at the top of files.